--- a/src/MySqlConnector/Constants.cs
+++ b/src/MySqlConnector/Constants.cs
@@ -1,0 +1,10 @@
+namespace MySql.Data
+{
+	internal static class Constants
+	{
+		/// <summary>
+		/// A sentinel value indicating no (or infinite) timeout.
+		/// </summary>
+		public const int InfiniteTimeout = int.MaxValue;
+	}
+}