--- a/src/MySqlConnector/Protocol/Serialization/SocketByteHandler.cs
+++ b/src/MySqlConnector/Protocol/Serialization/SocketByteHandler.cs
@@ -1,6 +1,7 @@
-﻿using System;
+using System;
 using System.Net.Sockets;
 using System.Threading.Tasks;
+using MySql.Data.MySqlClient;
 
 namespace MySql.Data.Protocol.Serialization
 {
@@ -11,21 +12,66 @@ namespace MySql.Data.Protocol.Serialization
 			m_socket = socket;
 			var socketEventArgs = new SocketAsyncEventArgs();
 			m_socketAwaitable = new SocketAwaitable(socketEventArgs);
+			m_closeSocket = socket.Dispose;
+			RemainingTimeout = Constants.InfiniteTimeout;
 		}
 
 		public void Dispose() => m_socketAwaitable.EventArgs.Dispose();
 
+		public int RemainingTimeout { get; set; }
+
 		public ValueTask<int> ReadBytesAsync(ArraySegment<byte> buffer, IOBehavior ioBehavior)
 		{
-			return (ioBehavior == IOBehavior.Asynchronous) ?
-				new ValueTask<int>(DoReadBytesAsync(buffer)) :
-				new ValueTask<int>(m_socket.Receive(buffer.Array, buffer.Offset, buffer.Count, SocketFlags.None));
+			return ioBehavior == IOBehavior.Asynchronous ?
+				new ValueTask<int>(DoReadBytesAsync(buffer)) : DoReadBytesSync(buffer);
+
+			ValueTask<int> DoReadBytesSync(ArraySegment<byte> buffer_)
+			{
+				if (RemainingTimeout == Constants.InfiniteTimeout)
+					return new ValueTask<int>(m_socket.Receive(buffer_.Array, buffer_.Offset, buffer_.Count, SocketFlags.None));
+
+				while (RemainingTimeout > 0)
+				{
+					var startTime = Environment.TickCount;
+					if (m_socket.Poll(Math.Min(int.MaxValue / 1000, RemainingTimeout) * 1000, SelectMode.SelectRead))
+					{
+						var bytesRead = m_socket.Receive(buffer_.Array, buffer_.Offset, buffer_.Count, SocketFlags.None);
+						RemainingTimeout -= unchecked(Environment.TickCount - startTime);
+						return new ValueTask<int>(bytesRead);
+					}
+					RemainingTimeout -= unchecked(Environment.TickCount - startTime);
+				}
+				return ValueTaskExtensions.FromException<int>(MySqlException.CreateForTimeout());
+			}
 
 			async Task<int> DoReadBytesAsync(ArraySegment<byte> buffer_)
 			{
+				var startTime = RemainingTimeout == Constants.InfiniteTimeout ? 0 : Environment.TickCount;
+				var timerId = RemainingTimeout == Constants.InfiniteTimeout ? 0 : TimerQueue.Instance.Add(RemainingTimeout, m_closeSocket);
 				m_socketAwaitable.EventArgs.SetBuffer(buffer_.Array, buffer_.Offset, buffer_.Count);
-				await m_socket.ReceiveAsync(m_socketAwaitable);
-				return m_socketAwaitable.EventArgs.BytesTransferred;
+				int bytesRead;
+				try
+				{
+					await m_socket.ReceiveAsync(m_socketAwaitable);
+					bytesRead =  m_socketAwaitable.EventArgs.BytesTransferred;
+				}
+				catch (SocketException ex)
+				{
+					if (RemainingTimeout != Constants.InfiniteTimeout)
+					{
+						RemainingTimeout -= unchecked(Environment.TickCount - startTime);
+						if (!TimerQueue.Instance.Remove(timerId))
+							throw MySqlException.CreateForTimeout(ex);
+					}
+					throw;
+				}
+				if (RemainingTimeout != Constants.InfiniteTimeout)
+				{
+					RemainingTimeout -= unchecked(Environment.TickCount - startTime);
+					if (!TimerQueue.Instance.Remove(timerId))
+						throw MySqlException.CreateForTimeout();
+				}
+				return bytesRead;
 			}
 		}
 
@@ -51,5 +97,6 @@ namespace MySql.Data.Protocol.Serialization
 
 		readonly Socket m_socket;
 		readonly SocketAwaitable m_socketAwaitable;
+		readonly Action m_closeSocket;
 	}
 }