--- a/src/MySqlConnector/Protocol/Serialization/StreamByteHandler.cs
+++ b/src/MySqlConnector/Protocol/Serialization/StreamByteHandler.cs
@@ -1,6 +1,8 @@
-﻿using System;
+using System;
 using System.IO;
+using System.Threading;
 using System.Threading.Tasks;
+using MySql.Data.MySqlClient;
 
 namespace MySql.Data.Protocol.Serialization
 {
@@ -9,18 +11,69 @@ namespace MySql.Data.Protocol.Serialization
 		public StreamByteHandler(Stream stream)
 		{
 			m_stream = stream;
+			m_closeStream = m_stream.Dispose;
+			RemainingTimeout = Constants.InfiniteTimeout;
 		}
 
 		public void Dispose() => m_stream.Dispose();
 
+		public int RemainingTimeout { get; set; }
+
 		public ValueTask<int> ReadBytesAsync(ArraySegment<byte> buffer, IOBehavior ioBehavior)
 		{
 			return (ioBehavior == IOBehavior.Asynchronous) ?
-				new ValueTask<int>(DoReadBytesAsync(buffer)) :
-				new ValueTask<int>(m_stream.Read(buffer.Array, buffer.Offset, buffer.Count));
+				new ValueTask<int>(DoReadBytesAsync(buffer)) : DoReadBytesSync(buffer);
 
-			async Task<int> DoReadBytesAsync(ArraySegment<byte> buffer_) =>
-				await m_stream.ReadAsync(buffer_.Array, buffer_.Offset, buffer_.Count).ConfigureAwait(false);
+			ValueTask<int> DoReadBytesSync(ArraySegment<byte> buffer_)
+			{
+				if (RemainingTimeout <= 0)
+					return ValueTaskExtensions.FromException<int>(MySqlException.CreateForTimeout());
+
+				m_stream.ReadTimeout = RemainingTimeout == Constants.InfiniteTimeout ? Timeout.Infinite : RemainingTimeout;
+				var startTime = RemainingTimeout == Constants.InfiniteTimeout ? 0 : Environment.TickCount;
+				int bytesRead;
+				try
+				{
+					bytesRead = m_stream.Read(buffer_.Array, buffer_.Offset, buffer_.Count);
+				}
+				catch (Exception ex)
+				{
+					if (ex is IOException && RemainingTimeout != Constants.InfiniteTimeout)
+						return ValueTaskExtensions.FromException<int>(MySqlException.CreateForTimeout(ex));
+					return ValueTaskExtensions.FromException<int>(ex);
+				}
+				if (RemainingTimeout != Constants.InfiniteTimeout)
+					RemainingTimeout -= unchecked(Environment.TickCount - startTime);
+				return new ValueTask<int>(bytesRead);
+			}
+
+			async Task<int> DoReadBytesAsync(ArraySegment<byte> buffer_)
+			{
+				var startTime = RemainingTimeout == Constants.InfiniteTimeout ? 0 : Environment.TickCount;
+				var timerId = RemainingTimeout == Constants.InfiniteTimeout ? 0 : TimerQueue.Instance.Add(RemainingTimeout, m_closeStream);
+				int bytesRead;
+				try
+				{
+					bytesRead = await m_stream.ReadAsync(buffer_.Array, buffer_.Offset, buffer_.Count).ConfigureAwait(false);
+				}
+				catch (ObjectDisposedException ex)
+				{
+					if (RemainingTimeout != Constants.InfiniteTimeout)
+					{
+						RemainingTimeout -= unchecked(Environment.TickCount - startTime);
+						if (!TimerQueue.Instance.Remove(timerId))
+							throw MySqlException.CreateForTimeout(ex);
+					}
+					throw;
+				}
+				if (RemainingTimeout != Constants.InfiniteTimeout)
+				{
+					RemainingTimeout -= unchecked(Environment.TickCount - startTime);
+					if (!TimerQueue.Instance.Remove(timerId))
+						throw MySqlException.CreateForTimeout();
+				}
+				return bytesRead;
+			}
 		}
 
 		public ValueTask<int> WriteBytesAsync(ArraySegment<byte> data, IOBehavior ioBehavior)
@@ -43,5 +96,6 @@ namespace MySql.Data.Protocol.Serialization
 		}
 
 		readonly Stream m_stream;
+		readonly Action m_closeStream;
 	}
 }