… .ci/config/config.buffer.json
… .ci/config/config.compression+ssl.json
… .ci/config/config.compression.json
… .ci/config/config.json
… .ci/config/config.ssl.json
… .ci/config/config.uds+ssl.json
… .ci/config/config.uds.json
… .travis.yml
… docs/content/connection-options.md
… docs/content/tutorials/migrating-from-connector-net.md
… tests/SideBySide/CommandTimeoutTests.cs
… tests/SideBySide/config.json.example