Adds `MySqlConnectionStringBuilder.DefaultCommandTimeout`. Respects `CommandTimeout` for synchronous and asynchronous Command/DataReader methods.