@bgrainger 2017-08-04
I've noticed a high number of failures in `SideBySide.CancelTests.CancelSlowQueryWithTokenAfterExecuteReader` after submitting this PR. I don't know if that's a side-effect of this new code, a coincidental environmental change in the CI environment, or something else.

It's probably worth eliminating as a new regression to ensure the `master` branch doesn't have spurious build breakages.