@caleblloyd 2017-08-04
> I've noticed a high number of failures in SideBySide.CancelTests.CancelSlowQueryWithTokenAfterExecuteReader after submitting this PR.

That's one that I've always had issues with in CI, I think we've updated the timeout significantly already.  I wonder if it's proportional to the amount of testing happening in general, like if it gets slower every time we add more tests