@bgrainger 2017-09-25
Failing tests:

Appveyor: `CommandTimeoutWithSleepSync` fails with `Compression, No SSL` inconsistently (`netcoreapp1.1.2` once, `net461` once).
Travis CI: `MultipleCommandTimeoutWithSleepSync` fails consistently under `netcoreapp2.0` (but passes under `netcoreapp1.1.2`).