@bgrainger 2017-10-19
# Benchmarks

## Before

 |             Method |       Job | Runtime |   Toolchain |        Mean |      Error |      StdDev |     StdErr |      Median |         Min |          Q1 |          Q3 |         Max |     Op/s |    Gen 0 |    Gen 1 |    Gen 2 |  Allocated |
 |------------------- |---------- |-------- |------------ |------------:|-----------:|------------:|-----------:|------------:|------------:|------------:|------------:|------------:|---------:|---------:|---------:|---------:|-----------:|
 |  OpenFromPoolAsync |     net47 |     Clr | CsProjnet47 |   215.52 us |  3.3467 us |   3.1305 us |  0.8083 us |   215.22 us |   211.75 us |   212.07 us |   217.96 us |   221.52 us |  4,639.9 |   1.4648 |        - |        - |    9.45 KB |
 |   OpenFromPoolSync |     net47 |     Clr | CsProjnet47 |    84.91 us |  1.6825 us |   1.5738 us |  0.4064 us |    84.53 us |    82.43 us |    84.12 us |    85.94 us |    88.50 us | 11,777.1 |   0.1221 |        - |        - |    1.43 KB |
 | ExecuteScalarAsync |     net47 |     Clr | CsProjnet47 |   117.70 us |  2.0712 us |   1.8361 us |  0.4907 us |   117.21 us |   115.60 us |   116.67 us |   117.95 us |   121.88 us |  8,496.0 |   0.9766 |        - |        - |    6.89 KB |
 |  ExecuteScalarSync |     net47 |     Clr | CsProjnet47 |    76.23 us |  1.4930 us |   2.1412 us |  0.4047 us |    75.87 us |    73.25 us |    74.42 us |    78.18 us |    80.34 us | 13,118.7 |   0.2441 |        - |        - |    2.05 KB |
 |     ReadBlobsAsync |     net47 |     Clr | CsProjnet47 | 2,017.27 us | 39.9334 us |  85.9610 us | 11.4870 us | 1,985.32 us | 1,922.94 us | 1,952.50 us | 2,075.94 us | 2,259.01 us |    495.7 | 332.0313 | 332.0313 | 332.0313 |  1106.1 KB |
 |      ReadBlobsSync |     net47 |     Clr | CsProjnet47 | 1,712.09 us | 33.3332 us |  31.1799 us |  8.0506 us | 1,701.62 us | 1,682.56 us | 1,689.88 us | 1,740.58 us | 1,785.70 us |    584.1 | 332.0313 | 332.0313 | 332.0313 | 1076.95 KB |
 |      ManyRowsAsync |     net47 |     Clr | CsProjnet47 | 5,722.13 us | 19.3930 us |  17.1913 us |  4.5946 us | 5,723.85 us | 5,698.87 us | 5,707.66 us | 5,731.99 us | 5,764.86 us |    174.8 |        - |        - |        - |    26.6 KB |
 |       ManyRowsSync |     net47 |     Clr | CsProjnet47 | 5,263.83 us | 32.8506 us |  29.1212 us |  7.7830 us | 5,255.17 us | 5,198.47 us | 5,249.72 us | 5,290.29 us | 5,312.66 us |    190.0 |        - |        - |        - |    3.19 KB |
 |  OpenFromPoolAsync | netcore11 |    Core |  CoreCsProj |   208.73 us |  1.2873 us |   1.2041 us |  0.3109 us |   208.60 us |   206.78 us |   207.90 us |   209.53 us |   210.78 us |  4,790.8 |   1.4648 |   0.2441 |        - |    3.38 KB |
 |   OpenFromPoolSync | netcore11 |    Core |  CoreCsProj |    80.32 us |  1.5893 us |   2.3788 us |  0.4343 us |    79.79 us |    77.44 us |    78.33 us |    81.00 us |    87.25 us | 12,449.9 |   0.1221 |        - |        - |    1.38 KB |
 | ExecuteScalarAsync | netcore11 |    Core |  CoreCsProj |   116.32 us |  2.3231 us |   5.9130 us |  0.6783 us |   113.92 us |   108.80 us |   111.69 us |   120.10 us |   132.22 us |  8,597.3 |   0.9766 |        - |        - |    2.75 KB |
 |  ExecuteScalarSync | netcore11 |    Core |  CoreCsProj |    72.07 us |  0.7953 us |   0.7440 us |  0.1921 us |    72.12 us |    70.99 us |    71.40 us |    72.50 us |    73.62 us | 13,876.1 |   0.2441 |        - |        - |    2.02 KB |
 |     ReadBlobsAsync | netcore11 |    Core |  CoreCsProj | 1,850.97 us | 35.8495 us |  74.0355 us | 10.2669 us | 1,841.01 us | 1,731.55 us | 1,788.24 us | 1,901.29 us | 2,023.10 us |    540.3 | 332.0313 | 332.0313 | 332.0313 |    2.54 KB |
 |      ReadBlobsSync | netcore11 |    Core |  CoreCsProj | 1,624.92 us |  8.8681 us |   8.2953 us |  2.1418 us | 1,630.30 us | 1,608.20 us | 1,616.80 us | 1,631.68 us | 1,633.00 us |    615.4 | 332.0313 | 332.0313 | 332.0313 | 1076.16 KB |
 |      ManyRowsAsync | netcore11 |    Core |  CoreCsProj | 5,009.96 us | 42.4936 us |  39.7486 us | 10.2630 us | 5,021.02 us | 4,923.31 us | 4,988.78 us | 5,035.97 us | 5,059.25 us |    199.6 |        - |        - |        - |    2.68 KB |
 |       ManyRowsSync | netcore11 |    Core |  CoreCsProj | 4,532.68 us | 29.0356 us |  27.1599 us |  7.0127 us | 4,533.30 us | 4,494.96 us | 4,510.41 us | 4,549.27 us | 4,586.02 us |    220.6 |        - |        - |        - |    3.16 KB |
 |  OpenFromPoolAsync | netcore20 |    Core |  CoreCsProj |   240.49 us |  9.5109 us |  28.0430 us |  2.8043 us |   233.02 us |   204.96 us |   215.66 us |   268.04 us |   300.88 us |  4,158.1 |   0.9766 |        - |        - |    3.73 KB |
 |   OpenFromPoolSync | netcore20 |    Core |  CoreCsProj |    81.88 us |  1.6063 us |   2.3037 us |  0.4354 us |    81.58 us |    78.84 us |    79.76 us |    83.61 us |    87.19 us | 12,212.7 |   0.1221 |        - |        - |    1.35 KB |
 | ExecuteScalarAsync | netcore20 |    Core |  CoreCsProj |   132.66 us |  3.0205 us |   8.9059 us |  0.8906 us |   132.97 us |   116.73 us |   123.99 us |   138.82 us |   156.47 us |  7,538.1 |   0.7324 |        - |        - |    4.17 KB |
 |  ExecuteScalarSync | netcore20 |    Core |  CoreCsProj |    71.32 us |  1.3762 us |   1.2873 us |  0.3324 us |    71.06 us |    70.03 us |    70.10 us |    72.20 us |    73.84 us | 14,021.2 |   0.2441 |        - |        - |    1.98 KB |
 |     ReadBlobsAsync | netcore20 |    Core |  CoreCsProj | 1,891.66 us | 38.7617 us |  93.6139 us | 11.2698 us | 1,853.36 us | 1,796.86 us | 1,829.50 us | 1,928.83 us | 2,193.42 us |    528.6 | 332.0313 | 332.0313 | 332.0313 |    4.08 KB |
 |      ReadBlobsSync | netcore20 |    Core |  CoreCsProj | 1,888.38 us | 86.0131 us | 241.1900 us | 25.2836 us | 1,782.67 us | 1,665.81 us | 1,722.92 us | 1,924.14 us | 2,661.30 us |    529.6 | 332.0313 | 332.0313 | 332.0313 | 1076.13 KB |
 |      ManyRowsAsync | netcore20 |    Core |  CoreCsProj | 4,047.57 us | 77.5274 us |  82.9534 us | 19.5523 us | 4,003.85 us | 3,967.96 us | 3,993.03 us | 4,078.50 us | 4,257.63 us |    247.1 |        - |        - |        - |    4.04 KB |
 |       ManyRowsSync | netcore20 |    Core |  CoreCsProj | 4,034.10 us | 80.1637 us |  74.9852 us | 19.3611 us | 4,007.19 us | 3,965.16 us | 3,979.47 us | 4,089.18 us | 4,241.85 us |    247.9 |        - |        - |        - |    3.05 KB |

## After

|             Method |       Job | Runtime |   Toolchain |        Mean |      Error |     StdDev |    StdErr |      Median |         Min |          Q1 |          Q3 |         Max |     Op/s |    Gen 0 |    Gen 1 |    Gen 2 |  Allocated |
 |------------------- |---------- |-------- |------------ |------------:|-----------:|-----------:|----------:|------------:|------------:|------------:|------------:|------------:|---------:|---------:|---------:|---------:|-----------:|
 |  OpenFromPoolAsync |     net47 |     Clr | CsProjnet47 |   214.80 us |  1.6467 us |  1.3751 us | 0.3814 us |   214.64 us |   212.41 us |   213.76 us |   215.69 us |   217.24 us |  4,655.5 |   1.4648 |        - |        - |    9.45 KB |
 |   OpenFromPoolSync |     net47 |     Clr | CsProjnet47 |    83.64 us |  0.9082 us |  0.8051 us | 0.2152 us |    83.44 us |    82.45 us |    83.06 us |    84.19 us |    85.19 us | 11,956.6 |   0.1221 |        - |        - |    1.43 KB |
 | ExecuteScalarAsync |     net47 |     Clr | CsProjnet47 |   120.18 us |  2.3184 us |  2.4807 us | 0.5847 us |   120.02 us |   116.81 us |   118.36 us |   122.05 us |   126.62 us |  8,321.1 |   0.9766 |        - |        - |     6.9 KB |
 |  ExecuteScalarSync |     net47 |     Clr | CsProjnet47 |    75.98 us |  0.5483 us |  0.5129 us | 0.1324 us |    75.85 us |    75.31 us |    75.55 us |    76.34 us |    77.32 us | 13,162.0 |   0.2441 |        - |        - |    2.09 KB |
 |     ReadBlobsAsync |     net47 |     Clr | CsProjnet47 | 1,986.04 us | 35.6083 us | 33.3080 us | 8.6001 us | 1,988.23 us | 1,931.23 us | 1,954.59 us | 2,007.34 us | 2,041.73 us |    503.5 | 332.0313 | 332.0313 | 332.0313 | 1106.12 KB |
 |      ReadBlobsSync |     net47 |     Clr | CsProjnet47 | 1,890.30 us | 18.0214 us | 15.9755 us | 4.2696 us | 1,892.29 us | 1,854.74 us | 1,883.36 us | 1,899.16 us | 1,918.71 us |    529.0 | 332.0313 | 332.0313 | 332.0313 | 1079.59 KB |
 |      ManyRowsAsync |     net47 |     Clr | CsProjnet47 | 5,956.84 us | 14.1689 us | 11.8316 us | 3.2815 us | 5,956.22 us | 5,934.22 us | 5,949.55 us | 5,964.17 us | 5,978.48 us |    167.9 |        - |        - |        - |   26.67 KB |
 |       ManyRowsSync |     net47 |     Clr | CsProjnet47 | 5,316.81 us | 25.6195 us | 22.7110 us | 6.0698 us | 5,314.65 us | 5,282.84 us | 5,300.51 us | 5,336.55 us | 5,357.98 us |    188.1 |        - |        - |        - |     3.5 KB |
 |  OpenFromPoolAsync | netcore11 |    Core |  CoreCsProj |   210.25 us |  0.4599 us |  0.4302 us | 0.1111 us |   210.23 us |   209.58 us |   209.91 us |   210.55 us |   211.12 us |  4,756.4 |   1.4648 |   0.2441 |        - |    3.38 KB |
 |   OpenFromPoolSync | netcore11 |    Core |  CoreCsProj |    83.72 us |  1.8118 us |  3.4472 us | 0.5139 us |    82.26 us |    79.82 us |    81.35 us |    85.25 us |    93.67 us | 11,944.8 |   0.1221 |        - |        - |    1.38 KB |
 | ExecuteScalarAsync | netcore11 |    Core |  CoreCsProj |   121.65 us |  2.3851 us |  4.1772 us | 0.6689 us |   120.79 us |   115.63 us |   118.10 us |   124.52 us |   130.33 us |  8,220.6 |   1.0986 |   0.1221 |        - |    2.75 KB |
 |  ExecuteScalarSync | netcore11 |    Core |  CoreCsProj |    75.05 us |  1.4750 us |  2.5047 us | 0.4118 us |    74.63 us |    71.81 us |    72.70 us |    76.86 us |    81.06 us | 13,324.8 |   0.2441 |        - |        - |    2.02 KB |
 |     ReadBlobsAsync | netcore11 |    Core |  CoreCsProj | 1,939.94 us | 32.1737 us | 30.0953 us | 7.7706 us | 1,928.82 us | 1,896.88 us | 1,919.79 us | 1,967.05 us | 1,989.47 us |    515.5 | 332.0313 | 332.0313 | 332.0313 |    2.54 KB |
 |      ReadBlobsSync | netcore11 |    Core |  CoreCsProj | 1,637.53 us | 33.0565 us | 29.3037 us | 7.8318 us | 1,624.30 us | 1,614.30 us | 1,619.53 us | 1,643.92 us | 1,715.51 us |    610.7 | 332.0313 | 332.0313 | 332.0313 | 1076.16 KB |
 |      ManyRowsAsync | netcore11 |    Core |  CoreCsProj | 5,036.84 us | 15.0494 us | 14.0772 us | 3.6347 us | 5,032.46 us | 5,017.68 us | 5,027.23 us | 5,048.34 us | 5,061.58 us |    198.5 |        - |        - |        - |    2.68 KB |
 |       ManyRowsSync | netcore11 |    Core |  CoreCsProj | 4,535.03 us | 15.8731 us | 12.3927 us | 3.5775 us | 4,537.96 us | 4,514.20 us | 4,525.20 us | 4,543.17 us | 4,552.40 us |    220.5 |        - |        - |        - |    3.16 KB |
 |  OpenFromPoolAsync | netcore20 |    Core |  CoreCsProj |   221.73 us |  4.3130 us |  4.9669 us | 1.1106 us |   219.95 us |   214.80 us |   218.71 us |   224.83 us |   234.61 us |  4,510.0 |   0.9766 |        - |        - |    3.73 KB |
 |   OpenFromPoolSync | netcore20 |    Core |  CoreCsProj |    83.31 us |  1.6589 us |  2.4830 us | 0.4533 us |    82.85 us |    80.19 us |    81.30 us |    85.18 us |    90.62 us | 12,003.0 |   0.1221 |        - |        - |    1.35 KB |
 | ExecuteScalarAsync | netcore20 |    Core |  CoreCsProj |   125.32 us |  2.6532 us |  6.0426 us | 0.7674 us |   124.55 us |   116.04 us |   120.82 us |   128.43 us |   144.96 us |  7,979.6 |   0.7324 |        - |        - |    4.17 KB |
 |  ExecuteScalarSync | netcore20 |    Core |  CoreCsProj |    73.09 us |  1.1260 us |  1.0533 us | 0.2720 us |    73.07 us |    71.78 us |    72.23 us |    74.01 us |    75.28 us | 13,681.6 |   0.2441 |        - |        - |    1.98 KB |
 |     ReadBlobsAsync | netcore20 |    Core |  CoreCsProj | 1,828.06 us | 36.4529 us | 55.6675 us | 9.9982 us | 1,808.48 us | 1,744.15 us | 1,782.86 us | 1,881.08 us | 1,950.24 us |    547.0 | 332.0313 | 332.0313 | 332.0313 |    3.97 KB |
 |      ReadBlobsSync | netcore20 |    Core |  CoreCsProj | 1,695.93 us | 36.8787 us | 43.9015 us | 9.5801 us | 1,677.71 us | 1,641.75 us | 1,666.23 us | 1,730.33 us | 1,799.90 us |    589.6 | 332.0313 | 332.0313 | 332.0313 | 1076.13 KB |
 |      ManyRowsAsync | netcore20 |    Core |  CoreCsProj | 4,006.36 us | 17.0521 us | 15.9505 us | 4.1184 us | 4,002.60 us | 3,978.97 us | 3,995.93 us | 4,017.68 us | 4,039.70 us |    249.6 |        - |        - |        - |    4.04 KB |
 |       ManyRowsSync | netcore20 |    Core |  CoreCsProj | 3,930.19 us | 14.9531 us | 12.4865 us | 3.4631 us | 3,928.85 us | 3,904.65 us | 3,921.47 us | 3,939.96 us | 3,949.22 us |    254.4 |        - |        - |        - |    3.05 KB |